+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## Unreleased
+
+## 14.0 (2020-04-22)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/13.0...14.0
+
+- 135 Employer logos (BE)
+- 176 Jobs fix import location
+- 204 Jobs fix salary imports (BE)
+
 ## 13.0 (2020-04-20)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/12.0...13.0
@@ -11,6 +21,7 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/12.0...13.0
 - Add favicon
 - 140 Job site breadcrumbs on all pages
 - 155 Jobs extra filters (BE)
+- 161 Update job search filter styles and add custom select form field
 - 173 Search local council area by postcode
 - 179 Jobs additional fields BE
 
@@ -22,6 +33,7 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/11.0...12.0
 - 172 Add live chat support feature
 - 174 Fix button padding
 - 175 Remove hard coded Bucks logo from job detail pages
+- 193 Fix issue with duplicate JobSubcategory titles
 
 ## 11.0 (2020-04-01)
 
@@ -46,7 +58,6 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/9.0...10.0
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
 
 - 129 Jobs postcode search
-- 161 Update job search filter styles and add custom select form field
 - 169 Adjust position of form field help text
 - Forms integration with Aptean Respond API
 - Fix an issue with CI deployments